Se realiza carga de datos para que se visualice en el catalogo

+++ app/models.py
@@ -37,7 +37,7 @@ class Producto(models.Model):
     id_producto = models.AutoField(primary_key=True)
     nombre_prod = models.CharField(max_length=100)
     descripcion = models.TextField()
-    valor = models.IntegerField(default=0) # valor por defecto 0
+    valor = models.DecimalField(max_digits=10, decimal_places=2, default=0.0) # valor por defecto 0
     color = models.CharField(max_length=50)
     id_categoria = models.ForeignKey(Categoria, on_delete=models.CASCADE)
     id_marca = models.ForeignKey(Marca, on_delete=models.CASCADE)
